refactor: introduce abstract PathDependency base class for FileDependency and DirectoryDependency
