+++ src/poetry/core/packages/file_dependency.py
@@ -4,14 +4,17 @@ import hashlib
 import io
 import warnings
 
-from pathlib import Path
+from typing import TYPE_CHECKING
 from typing import Iterable
 
-from poetry.core.packages.dependency import Dependency
-from poetry.core.packages.utils.utils import path_to_url
+from poetry.core.packages.path_dependency import PathDependency
 
 
-class FileDependency(Dependency):
+if TYPE_CHECKING:
+    from pathlib import Path
+
+
+class FileDependency(PathDependency):
     def __init__(
         self,
         name: str,
@@ -21,47 +24,18 @@ class FileDependency(Dependency):
         base: Path | None = None,
         extras: Iterable[str] | None = None,
     ) -> None:
-        self._path = path
-        self._base = base or Path.cwd()
-        self._full_path = path
-
-        if not self._path.is_absolute():
-            try:
-                self._full_path = self._base.joinpath(self._path).resolve()
-            except FileNotFoundError:
-                raise ValueError(f"Directory {self._path} does not exist")
-
-        if not self._full_path.exists():
-            raise ValueError(f"File {self._path} does not exist")
-
-        if self._full_path.is_dir():
-            raise ValueError(f"{self._path} is a directory, expected a file")
-
         super().__init__(
             name,
-            "*",
+            path,
+            source_type="file",
             groups=groups,
             optional=optional,
-            allows_prereleases=True,
-            source_type="file",
-            source_url=self._full_path.as_posix(),
+            base=base,
             extras=extras,
         )
 
-    @property
-    def base(self) -> Path:
-        return self._base
-
-    @property
-    def path(self) -> Path:
-        return self._path
-
-    @property
-    def full_path(self) -> Path:
-        return self._full_path
-
-    def is_file(self) -> bool:
-        return True
+        if self._full_path.is_dir():
+            raise ValueError(f"{self._path} is a directory, expected a file")
 
     def hash(self, hash_name: str = "sha256") -> str:
         warnings.warn(
@@ -75,16 +49,3 @@ class FileDependency(Dependency):
                 h.update(content)
 
         return h.hexdigest()
-
-    @property
-    def base_pep_508_name(self) -> str:
-        requirement = self.pretty_name
-
-        if self.extras:
-            extras = ",".join(sorted(self.extras))
-            requirement += f"[{extras}]"
-
-        path = path_to_url(self.full_path)
-        requirement += f" @ {path}"
-
-        return requirement
